+++ src/helpers.ts
@@ -76,3 +76,10 @@ export function pathToJestPackageJSON(pluginSettings: IPluginSettings): string |
 function removeSurroundingQuotes(str) {
   return str.replace(/^['"`]/, '').replace(/['"`]$/, '')
 }
+
+/**
+ *  Taken From https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions
+ */
+export function escapeRegExp(str: string): string {
+  return str.replace(/[.*+?^${}()|[\]\\]/g, '\\$&') // $& means the whole matched string
+}